requirements.txt
numpy>=1.24
pandas>=2.0
python-dateutil>=2.8

# PDF reading (optional, only needed if you pass .pdf files)
pypdf>=4.0

# Gemini AI mapping/suggestions (optional, only if you use --map_ai or --ai)
google-genai>=0.5.0
